requirements.txt
google-generativeai==0.3.2
flask==2.3.3
python-dotenv==1.0.0
gunicorn==21.2.0 